Big refactor; Multiple nics (interfaces); IPv4Address endianess bugfixes; Improved NetworkManager; DHCP & DNS working; DNS recursion implemented; Fixed bugs in NameParser; Shorter commands; Added testcase in README.

+++ src/io/LowlevelOutput.java
@@ -50,7 +50,7 @@ public class LowlevelOutput {
 
         char tmp;
         for (int i = len - 1; i >= 1; i--) {
-            printChar(alphabet.charAt((int) (value % base)), x + i, y, color);
+            printChar(alphabet.charAt(value % base), x + i, y, color);
             value /= base;
         }
